nix/ci/execline: add nm and strip to cargoEnvironment on Darwin
On Darwin, some tools are not provided by nix, but have to be taken
from /usr/bin in the environment. Since we cleaned the environment, we
add them to PATH again. We don’t have to add everything in /usr/bin
however, we can create a symlink farm with just the tools we need.