Move CONTRIBUTING.md to toplevel
This makes CONTRIBUTING.md more prominent - people who want to
contribute are more likely to find this file in the toplevel.

According to the Github help, the Github UI picks up CONTRIBUTING.md
from the toplevel or from the .github folder, so this will not impact
this file's visibility via the Github UI (the thing that says, "It looks
like this is your first contribution to this project - have you had a
look at CONTRIBUTING.md yet?" or something like that).

https://help.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors